Split the interrupt tests out of the snapshot-run suite
The suite had grown past the ~400-line target, so the six tests about
signals and interrupts now live in their own interrupts.test.ts. Both
files sit well under the limit and mutation mapping stays narrow.
